+++ Demo/Demo Base/BraintreeDemoIntegrationViewController.m
@@ -3,7 +3,8 @@
 #import <InAppSettingsKit/IASKSpecifierValuesViewController.h>
 #import <InAppSettingsKit/IASKViewController.h>
 #import <InAppSettingsKit/IASKSettingsReader.h>
-#import <iOS-Slide-Menu/SlideNavigationController.h>
+
+@import iOS_Slide_Menu;
 
 @interface BraintreeDemoIntegrationViewController ()
 @property (nonatomic, strong) IASKSpecifierValuesViewController *targetViewController;