+++ UnitTests/BTConfiguration_Tests.swift
@@ -1,17 +1,7 @@
 import XCTest
 import PassKit
 
-#if swift(>=3.1)
-// Warning
-#else
-import VisaCheckoutSDK
-#endif
-
 class BTConfiguration_Tests: XCTestCase {
-
-    override func tearDown() {
-        BTConfiguration.setBetaPaymentOption("venmo", isEnabled: false)
-    }
 
     // MARK: - Visa Checkout category methods
 
@@ -42,9 +32,6 @@ class BTConfiguration_Tests: XCTestCase {
         XCTAssertFalse(configuration.isVisaCheckoutEnabled)
     }
 
-    #if swift(>=3.1)
-    // Warning
-    #else
     func testVisaCheckoutSupportedNetworks_returnsSupportedNetworks() {
         let configurationJSON = BTJSON(value: [
             "visaCheckout": [
@@ -78,7 +65,6 @@ class BTConfiguration_Tests: XCTestCase {
 
         XCTAssertEqual(configuration.visaCheckoutSupportedNetworks, [CardBrand.visa.rawValue, CardBrand.mastercard.rawValue, CardBrand.amex.rawValue, CardBrand.discover.rawValue])
     }
-    #endif
 
     func testVisaCheckoutSupportedNetworks_whenSupportedCardTypesIsMissing_returnsEmptyList() {
         let configurationJSON = BTJSON(value: [:])