+++ UnitTests/UnitTests-Bridging-Header.h
@@ -1,0 +1,12 @@
+//
+//  Use this file to import your target's public headers that you would like to expose to Swift.
+//
+
+@import Braintree;
+#import "BTConfiguration+VisaCheckout.h"
+#import "BTVisaCheckoutAddress.h"
+#import "BTVisaCheckoutCardNonce.h"
+#import "BTVisaCheckoutUserData.h"
+
+// Internal headers for testing
+#import "BTVisaCheckoutClient_Internal.h"